[core] Move index.ts to src/index.ts
